+++ test/SILOptimizer/exclusivity_static_diagnostics.swift
@@ -4,26 +4,6 @@ import Swift
 
 func takesTwoInouts<T>(_ p1: inout T, _ p2: inout T) { }
 
-func simpleInoutDiagnostic() {
-  var i = 7
-
-  // FIXME: This diagnostic should be removed if static enforcement is
-  // turned on by default.
-  // expected-error@+4{{inout arguments are not allowed to alias each other}}
-  // expected-note@+3{{previous aliasing argument}}
-  // expected-error@+2{{overlapping accesses to 'i', but modification requires exclusive access; consider copying to a local variable}}
-  // expected-note@+1{{conflicting access is here}}
-  takesTwoInouts(&i, &i)
-}
-
-func inoutOnInoutParameter(p: inout Int) {
-  // expected-error@+4{{inout arguments are not allowed to alias each other}}
-  // expected-note@+3{{previous aliasing argument}}
-  // expected-error@+2{{overlapping accesses to 'p', but modification requires exclusive access; consider copying to a local variable}}
-  // expected-note@+1{{conflicting access is here}}
-  takesTwoInouts(&p, &p)
-}
-
 func swapNoSuppression(_ i: Int, _ j: Int) {
   var a: [Int] = [1, 2, 3]
 
@@ -38,14 +18,6 @@ struct StructWithMutatingMethodThatTakesSelfInout {
   var f = SomeClass()
   mutating func mutate(_ other: inout StructWithMutatingMethodThatTakesSelfInout) { }
   mutating func mutate(_ other: inout SomeClass) { }
-
-  mutating func callMutatingMethodThatTakesSelfInout() {
-    // expected-error@+4{{inout arguments are not allowed to alias each other}}
-    // expected-note@+3{{previous aliasing argument}}
-    // expected-error@+2{{overlapping accesses to 'self', but modification requires exclusive access; consider copying to a local variable}}
-    // expected-note@+1{{conflicting access is here}}
-    mutate(&self)
-  }
 
   mutating func callMutatingMethodThatTakesSelfStoredPropInout() {
     // expected-error@+2{{overlapping accesses to 'self', but modification requires exclusive access; consider copying to a local variable}}
@@ -80,16 +52,6 @@ class ClassWithFinalStoredProp {
     local1.s1.mutate(&local1.s1.f)
   }
 }
-
-func violationWithGenericType<T>(_ p: T) {
-  var local = p
-  // expected-error@+4{{inout arguments are not allowed to alias each other}}
-  // expected-note@+3{{previous aliasing argument}}
-  // expected-error@+2{{overlapping accesses to 'local', but modification requires exclusive access; consider copying to a local variable}}
-  // expected-note@+1{{conflicting access is here}}
-  takesTwoInouts(&local, &local)
-}
-
 
 // Helper.
 struct StructWithTwoStoredProp {
